use pandas dataframes as common format for timeseries
these allow easy reading and writing.
here i did:
* implemented a simple writing routine.
* deleted utility.py as all functions were out of use.
